accept any value for options with required argument
This resolves #136

Only for options with optional arguments we have to check if the next argument
is an argument or another option. This allows to start arguments with a hyphen
(or dash `-`) like GNUs `getopt`. For optional arguments starting with a hyphen
you still have to provide them within the argument `--foo=-1`.